requirements.txt
langchain_community
langchain_text_splitters
langchain_openai
langchain_ollama
langchain_chroma
gradio
python-dotenv
pypdf
streamlit
google-genai